@@ -3,17 +3,16 @@
   "version": "0.1.0",
   "private": true,
   "devDependencies": {
-    "@types/node": "^16.10.3",
-    "@types/react": "^17.0.27",
+    "@types/node": "^16.7.13",
+    "@types/react": "^17.0.20",
     "@types/react-dom": "^17.0.9",
-    "@types/react-router-dom": "^5.3.0",
-    "react-scripts": "4.0.3",
-    "typescript": "^4.1.2"
+    "react-scripts": "5.0.0",
+    "typescript": "^4.4.2"
   },
   "dependencies": {
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-router-dom": "^5.3.0"
+    "react-router-dom": "^6.2.1"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -36,5 +35,5 @@
       "last 1 safari version"
     ]
   },
-  "proxy": "http://localhost:8080/"
+  "proxy": "http://localhost/"
 }
