@@ -1,5 +1,5 @@
 import React, { FunctionComponent, useEffect, useState } from 'react';
-import { useHistory } from 'react-router-dom';
+import { useNavigate } from 'react-router-dom';
 
 export interface AuthContextValue {
     isLoggedIn: boolean,
@@ -18,7 +18,7 @@ const AuthContext = React.createContext<AuthContextValue>({
 export const AuthContextProvider: FunctionComponent = (props) => {
     const [isLoggedIn, setIsLoggedIn] = useState(false);
 
-    const history = useHistory();
+    const navigate = useNavigate();
 
     useEffect(() => {
         const storedUserLoggedInInformation = localStorage.getItem('isLoggedIn');
@@ -33,14 +33,14 @@ export const AuthContextProvider: FunctionComponent = (props) => {
         // But it's just a dummy/ demo anyways
         localStorage.setItem('isLoggedIn', '1');
         setIsLoggedIn(true);
-        history.push('/');
+        navigate('/');
     };
 
     const logout = () => {
         console.log('AuthContextProvider.logout');
         localStorage.removeItem('isLoggedIn');
         setIsLoggedIn(false);
-        history.push('/');
+        navigate('/');
     };
 
     return (
