@@ -1,46 +1,20 @@
 import React, { FunctionComponent } from 'react';
 
 import { AuthContextProvider } from './context/auth-context';
+
 import Layout from './component/layout/Layout';
-import { BrowserRouter as Router, Redirect, Route, Switch } from 'react-router-dom';
-import HomePage from './page/HomePage';
-import NotFoundPage from './page/NotFoundPage';
-import TodoDetailPage from './page/TodoDetailPage';
-import NewTodoPage from './page/NewTodoPage';
-import LoginPage from './page/LoginPage';
-import SignUpPage from './page/SignUpPage';
+import { BrowserRouter } from 'react-router-dom';
+import AppRoutes from './AppRoutes';
 
 const App: FunctionComponent = () => {
     return (
-        <Router>
+        <BrowserRouter>
             <AuthContextProvider>
                 <Layout>
-                    <Switch>
-                        <Route path="/" exact>
-                            <Redirect to="/todos"/>
-                        </Route>
-                        <Route path="/todos" exact>
-                            <HomePage/>
-                        </Route>
-                        <Route path="/todos/:todoId">
-                            <TodoDetailPage/>
-                        </Route>
-                        <Route path="/new-todo">
-                            <NewTodoPage/>
-                        </Route>
-                        <Route path="/log-in">
-                            <LoginPage/>
-                        </Route>
-                        <Route path="/sign-up">
-                            <SignUpPage/>
-                        </Route>
-                        <Route path="*">
-                            <NotFoundPage/>
-                        </Route>
-                    </Switch>
+                    <AppRoutes/>
                 </Layout>
             </AuthContextProvider>
-        </Router>
+        </BrowserRouter>
     );
 }
 
