@@ -4,8 +4,6 @@ import './index.css';
 import App from './App';
 
 ReactDOM.render(
-    <React.StrictMode>
-        <App/>
-    </React.StrictMode>,
+    <App/>,
     document.getElementById('root')
 );
